Fixed bounds on scalar operations and added `Scalar` reexport.
Usually zero-dimensional tensors (shape of []) represent scalars, and one-dimensional tensors of size one (shape of [1]) are not scalars. This is also the way TensorFlow represents the shapes of scalars.

Another point is the necessity for the `Scalar` newtype. As of now I don't see any cases where a result of type `Scalar a` doesn't immediately gets unwrapped into type `a`. Maybe we could stop providing Runnable instance into `Scalar a` replacing it with `a`.